Move SymbolLoader to top level
