+++ lib/tapioca/internal.rb
@@ -15,7 +15,7 @@ require "tapioca/cli"
 require "tapioca/gemfile"
 require "tapioca/executor"
 require "tapioca/symbol_table_parser"
+require "tapioca/symbol_loader"
 require "tapioca/compilers/requires_compiler"
 require "tapioca/compilers/symbol_table_compiler"
-require "tapioca/compilers/symbol_table/symbol_loader"
 require "tapioca/compilers/dsl_compiler"
